Added undo transactions, to support save all and delete all.  All operations in a transaction will be undone when calling Persister.undo(), or redone when calling Persister.redo().  By default, calling Saveable.save or Saveable.delete will wrap the single generated operation in a transaction.  And Saveable.saveAll and Saveable.deleteAll will wrap all of the generated operations into a single transaction.
deleteAll requires overriding deleteRelated, in order to delete objects related to the Saveable object.  In general, the objects saved in saveRelated should be deleted in deleteRelated.

Have not yet experimented or tested using withUndoTransaction for other kinds of operations beyond saveAll And deleteAll.  withUndoTransaction also wraps all the DB operations executed in the passed in closure in a SQLite transaction.  So if any of those operations fail, they will all be rolled back.